For the love of God if this isn't the last commit for refactoring I will explode
