+++ MusiQueue-android/app/src/main/java/com/example/musiqueue/Activities/PlayerActivity.java
@@ -2,6 +2,7 @@ package com.example.musiqueue.Activities;
 import android.os.Bundle;
 import android.widget.Toast;
 
+import com.example.musiqueue.HelperClasses.YoutubeConnector;
 import com.example.musiqueue.R;
 import com.google.android.youtube.player.YouTubeBaseActivity;
 import com.google.android.youtube.player.YouTubeInitializationResult;
