+++ MusiQueue-android/app/src/main/java/com/example/musiqueue/Activities/hub_admin_only/QueueActivity.java
@@ -24,15 +24,16 @@ import android.widget.TextView;
 import android.widget.Toast;
 import android.widget.ToggleButton;
 
-import com.example.musiqueue.Activities.BackgroundWorker;
-import com.example.musiqueue.Activities.PollData;
-import com.example.musiqueue.Activities.UpdateResultReceiver;
-import com.example.musiqueue.Activities.UserItemAdapter;
-import com.example.musiqueue.Activities.VideoItem;
-import com.example.musiqueue.Activities.VideoItemAdapter;
-import com.example.musiqueue.Activities.YoutubeConnector;
 import com.example.musiqueue.Activities.users_only.QueueSong;
+import com.example.musiqueue.HelperClasses.BackgroundWorker;
 import com.example.musiqueue.HelperClasses.HubSingleton;
+import com.example.musiqueue.HelperClasses.PollData;
+import com.example.musiqueue.HelperClasses.UpdateResultReceiver;
+import com.example.musiqueue.HelperClasses.User;
+import com.example.musiqueue.HelperClasses.VideoItem;
+import com.example.musiqueue.HelperClasses.YoutubeConnector;
+import com.example.musiqueue.HelperClasses.adapters.UserItemAdapter;
+import com.example.musiqueue.HelperClasses.adapters.VideoItemAdapter;
 import com.example.musiqueue.R;
 import com.google.android.youtube.player.YouTubeInitializationResult;
 import com.google.android.youtube.player.YouTubePlayer;
