+++ MusiQueue-android/app/src/main/java/com/example/musiqueue/Activities/users_only/SearchHubsFragment.java
@@ -15,11 +15,11 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
 
-import com.example.musiqueue.Activities.BackgroundWorker;
+import com.example.musiqueue.HelperClasses.BackgroundWorker;
 import com.example.musiqueue.HelperClasses.HubSingleton;
-import com.example.musiqueue.HelperClasses.HubsListAdapter;
 import com.example.musiqueue.HelperClasses.HubsListItem;
 import com.example.musiqueue.HelperClasses.SearchHubResponse;
+import com.example.musiqueue.HelperClasses.adapters.HubsListAdapter;
 import com.example.musiqueue.R;
 import com.google.gson.Gson;
