+++ MusiQueue-android/app/src/main/java/com/example/musiqueue/HelperClasses/HubSingleton.java
@@ -1,6 +1,5 @@
 package com.example.musiqueue.HelperClasses;
 
-import com.example.musiqueue.Activities.hub_admin_only.User;
 import com.example.musiqueue.Activities.users_only.QueueSong;
 
 import java.util.ArrayList;
