+++ MusiQueue-android/app/src/main/java/com/example/musiqueue/HelperClasses/UpdateResultReceiver.java
@@ -1,4 +1,4 @@
-package com.example.musiqueue.Activities;
+package com.example.musiqueue.HelperClasses;
 
 import android.os.Bundle;
 import android.os.Handler;
