+++ MusiQueue-android/app/src/main/java/com/example/musiqueue/HelperClasses/WiFiBroadcastReceiver.java
@@ -1,4 +1,4 @@
-package com.example.musiqueue.Activities;
+package com.example.musiqueue.HelperClasses;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
