+++ MusiQueue-android/app/src/main/java/com/example/musiqueue/HelperClasses/adapters/HubsListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.musiqueue.HelperClasses;
+package com.example.musiqueue.HelperClasses.adapters;
 
 import android.content.Context;
 import android.support.v7.widget.RecyclerView;
@@ -7,6 +7,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import com.example.musiqueue.HelperClasses.HubsListItem;
 import com.example.musiqueue.R;
 
 import java.util.Vector;
